do-core1: Move Instruction to the do-core crate
