+++ do-core/src/instruction.rs
@@ -1,4 +1,4 @@
-use crate::Error;
+use crate::{Error, MAX_REGISTER_INDEX};
 
 #[allow(dead_code)]
 #[derive(Clone, Debug, PartialEq)]
@@ -18,5 +18,31 @@ impl OpCode {
             0x03 => Ok(OpCode::XOR),
             _ => Err(Error::InvalidOpCode(opcode)),
         }
+    }
+}
+
+#[derive(Debug)]
+pub struct Instruction {
+    opcode: OpCode,
+    op0: u8,
+    op1: u8,
+}
+
+impl Instruction {
+    // Instruction constructor, a.k.a. disassembler.
+    pub fn disassemble(insn: u16) -> Result<Instruction, Error> {
+        let opcode = OpCode::from_u8((insn >> 8) as u8)?;
+        let op0 = ((insn & 0xf0) >> 4) as u8;
+        let op1: u8 = (insn & 0xf) as u8;
+
+        if op0 > MAX_REGISTER_INDEX {
+            return Err(Error::Op0OutOfRange);
+        }
+
+        if op1 > MAX_REGISTER_INDEX {
+            return Err(Error::Op1OutOfRange);
+        }
+
+        Ok(Instruction { opcode, op0, op1 })
     }
 }
