+++ do-core/src/lib.rs
@@ -6,4 +6,7 @@ pub enum Error {
     AdditionOverflow(u16, u16),
 }
 
+// do-core register indexes range from 0 to 7.
+pub const MAX_REGISTER_INDEX: u8 = 7;
+
 pub mod instruction;
